Fix SEGFAULT on received error on server
This was caused by the fact that the error code is received as an
integer (e.g. 3003), and subsequently used as an index for the
error code arrays. However, these arrays naturally did not stretch
that far, thus causing a SEGFAULT.

In addition, the server does not skip straight to EXIT anymore once
it decodes a non-zero error code. Instead, it actually goes through
the message processing switch case and sends a FAILURE to the peer.
Thus the peer will know to stop, instead of sending the same message
over and over again until it hits its limit.