Gracefully halt when peer can not identify SSID
Previously, the peer would SEGFAULT at the moment it wanted to write
to the database to update the initial exchange info but it discovered
that there was no SSID set.

The peer uses the SSID to uniquely identify EAP-NOOB associations.
Therefore it is required to have one set such that EAP-NOOB works
correctly.